Create two udp senders
Instantiate a second copy of the udp sender using another wire pair.
This didn't work with the netgear hub; we suspect it couldn't deal
collisions from both ports active at the same time. We found a switch
that could recognize the packets (a Tenda wifi router) and the
receiver is able to see double the throughput (~20 MBps). However
it gets stuck now and then and packets stop getting through.
Changing the rate gets it unstuck for a while. The receiver is a
raspberry pi 3B.

We could take this further and connect the wifi router to the network
and see if the udp packets can be routed to other machines.