Simple demo accessing a HM628128-7 128Kx8 static ram.
The demo is a pointer chase in 256 bytes. In pseudocode:

  // Initialization
  for (i=0; i<256; i++)
    mem[i] = (i+1) & 0xff;

  // Test
  i = 0;
  while (true)
    i = mem[i];

The pin mapping is for MAXII EPM240 (CPLD) on a board with pin numbers
that correspond to the chip pins.

The SRAM's A8-A16 should be tied to ground. The chip selects should also
be tied: CS2 to Vcc and CS1_ to ground.

A0-A7 may be connected to LEDs to watch the progress of the chase.
There is a bug where A0 is stuck high.