TMS4464 64K x 4 DRAM demo driver
Driver with a simple memtest program. The driver is implemented
with microcode, in particular because the delays between steps
cause bloat in a state machine - maybe there's a simpler way?

The memtest writes an incrementing pattern to all of memory, then
reads it back forever. The "error" LED is set if an mismatch occurs.

Unexpectedly the row address has to be set up 2 cycles before RAS#,
which is not how I read the datasheet. To be investigated.

We tried to connect a MAX7219 for debugging but that's somehow not
working. Instead we used a logic analyzer.