[simm_test] Demo of interfacing to a 30-pin SIMM (16MB)
The test fills 16MB with a byte pattern and checks it.

(30-pin simms only go up to 16MB due to having 12 address and 8 data pins.)
This was the model (we don't use the parity bit):
https://keystron.com/collections/computer-memory/products/16mb-30pin-simm-ram-memory-with-parity-16x9-60ns-for-apple-macintosh-musical-sampler-old-pc-video-controller

We peeled the sticker off (warranty voided, lol) to look up the
datasheet for the DRAM. Surprisingly it uses a pair of Micron MT4LC16M4T8
16M x 4 chips which are specified for 3.3v. It is not 5v tolerant.
Absolute max is 4v. Well, I had started by feeding them 5v since that's
what the SIMM spec on wikipedia says. And I wonder how this functions
in the old 5v systems that seem to be its intended market.

But I have since confirmed that the SIMM works at 3v3. It's connected
directly to the FPGA using LVTTL 3.3.  So we can run it in spec and without
level shifters, how glorious.