[dram] Pause when toggling BB updown count. More max7219 tweaking.
Pause for a second when BB finishes so the counter can be seen. (It's
not an issue on the 16MB SIMM since zeroing it naturally takes a while).

More endless tweaking of the max7219 to try and keep it alive.
This change runs the display-off step exactly once, at reset.
(If we don't execute the display-off step at all, the display can get
trashed stuck even with power cycling. However if we include it in the
reset that is done periodically to recover from *whatever*, that seems
responsible for the display turning off.

Really would like to solve this but I can't figure what is going on.
Crappy devices or noise on the board or verilog bugs or low voltage...