mbv demo on small artix-7 with bram over axi and busybeaver
Current goal is to add SPI serial ram as RAM expansion for microblaze.
This will expand the qmtech xc7a15t memory from 64K BRAM to a few MB
which gives breathing room for programming in C/C++.

We are currently trying to implement the core for serial ram and
attach it to AXI.  For the AXI connection we thought to start with the
AXI full IP example but could not get it to work reliably with
microblaze even with plain bram.  Maybe the core is sending funky AXI
requests with bursts and partial words that the interface glosses
over. We'd need to debug with a testbench or ILA and learn AXI
details. For now we choose to try a shortcut.

The shortcut is to first use an AXI bram controller and bram generator
pair with a simple connection between them (addr, data and direction
signals). This does work with microblaze. The bram controller memory
latency must be fixed at synthesis but can be configured up to 128
cycles. Serial ram is deterministic so that should work e.g. a 32-bit
access takes 64 SPI cycles in standard mode or 16 SPI cycles with
QPI. (The current design includes AXI SPI connected to a microchip
512KB device with C code issuing manual commands.  It was unreliable
even under 10MHz on a breadboard but wirewrapping an SMD breakout
board or 8-pin dip to the gpio pins is a way forward.)

Next step is to replace the block ram generator with which the bram
controller normally pairs with our own block ram implementation. Then
add latency artifically to simulate serial ram and preview the
performance. We would not need to work with AXI to turn this into an
SPI interface. The less complexity the better while we deal in turn
with SPI, QSPI and tuning clock speed.